// crudapp2.0/src/app/model/movie.ts
// export class Movie{

    
//     movieName:string|undefined;
//     movieDirector:string|undefined;
//     movieGenre:string|undefined;
//     movieReleaseDate:string|undefined;
//     movieLanguage:string|undefined;
//     duration:string|undefined;
//     country:string|undefined;
//     description:string|undefined;
//     overallRate:number|undefined;
//     imageFile :File|undefined;


// }


export class Movie {
    movieName: string = '';
    movieDirector: string = '';
    movieGenre: string = '';
    movieReleaseDate: string = '';
    movieLanguage: string = '';
    duration: string = '';
    country: string = '';
    description: string = '';
    overallRate: number = 0;
    imageFile: File | undefined;
  }
  